Moved bodySiteTemplate into snippet.
